Many more implicit conversion bugs...
I haven't tested all these obviously but they fit the pattern and look
safe.

Still not done: cases that show up or propagate into qsortFun()/
iqsortSFun()/KnuthMorrisPratt() (statics which show up in multiple
places), many in Vector_* because I'm not sure they have the same
issue.

There are a lot of these.  Perhaps Range_index() and Array should
return/use I32 and COS be documented to only support Arrays of
~INT32_MAX (maybe -1 or something), which is effectively the case for
much of the API anyway.  But that's a breaking change to existing simple
uses of large arrays.